Use sets to track extractable nodes in extractor
Using integers to track number of desired nodes
extractable was a bug. It led to double-counting.
If this extractor is too slow in the future, we
could make it more efficient by using a more
efficient set; we don't need a HashSet, we could
simply use a bitvector which tracks inclusion
using a 0 or a 1.